+++ Jazz Jack Rabbit 2/RectCollider.cpp
@@ -6,15 +6,15 @@ void RectCollider::GiveDamageObject1(GameObject* object1, int sizeX1, int sizeY1
 	object1->GetPos();
 	object2->GetPos();
 
-	rect2.left = object1->GetPos().x - sizeX1 / 2;
-	rect2.right = object1->GetPos().x + sizeX1 / 2;
-	rect2.top = object1->GetPos().y - sizeY1;
-	rect2.bottom = object1->GetPos().y;
+	rect2.left = (LONG)(object1->GetPos().x - sizeX1 / 2);
+	rect2.right = (LONG)(object1->GetPos().x + sizeX1 / 2);
+	rect2.top = (LONG)(object1->GetPos().y - sizeY1);
+	rect2.bottom = (LONG)object1->GetPos().y;
 
-	rect3.left = object2->GetPos().x - sizeX2 / 2;
-	rect3.right = object2->GetPos().x + sizeX2 / 2;
-	rect3.top = object2->GetPos().y - sizeY2;
-	rect3.bottom = object2->GetPos().y;
+	rect3.left = (LONG)(object2->GetPos().x - sizeX2 / 2);
+	rect3.right = (LONG)(object2->GetPos().x + sizeX2 / 2);
+	rect3.top = (LONG)(object2->GetPos().y - sizeY2);
+	rect3.bottom = (LONG)object2->GetPos().y;
 
 	if (IntersectRect(&rect1, &rect2, &rect3))
 	{
@@ -27,15 +27,15 @@ void RectCollider::CollideObjects(GameObject* object1, int sizeX1, int sizeY1, G
 	object1->GetPos();
 	object2->GetPos();
 
-	rect2.left = object1->GetPos().x - sizeX1 / 2;
-	rect2.right = object1->GetPos().x + sizeX1 / 2;
-	rect2.top = object1->GetPos().y - sizeY1;
-	rect2.bottom = object1->GetPos().y;
+	rect2.left = (LONG)(object1->GetPos().x - sizeX1 / 2);
+	rect2.right = (LONG)(object1->GetPos().x + sizeX1 / 2);
+	rect2.top = (LONG)(object1->GetPos().y - sizeY1);
+	rect2.bottom = (LONG)object1->GetPos().y;
 
-	rect3.left = object2->GetPos().x - sizeX2 / 2;
-	rect3.right = object2->GetPos().x + sizeX2 / 2;
-	rect3.top = object2->GetPos().y - sizeY2;
-	rect3.bottom = object2->GetPos().y;
+	rect3.left = (LONG)(object2->GetPos().x - sizeX2 / 2);
+	rect3.right = (LONG)(object2->GetPos().x + sizeX2 / 2);
+	rect3.top = (LONG)(object2->GetPos().y - sizeY2);
+	rect3.bottom = (LONG)object2->GetPos().y;
 
 	if (IntersectRect(&rect1, &rect2, &rect3))
 	{
